Add tests for filter log
